requirements.txt
django-model-utils>=2.0
djangorestframework>=3.9.4

# Additional requirements go here
